requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
python-multipart==0.0.6
alembic==1.12.1
uvicorn
sqlalchemy[asyncio]
aiomysql
httpx
pydantic
python-dotenv
pytz==2023.3
requests==2.31.0